SKILL-177: add write-time test-value discipline directive for plan and mutating phases.
